+++ app/includes/content/general/contact.php
@@ -1,0 +1,8 @@
+<!--stories-->
+<div class="row" id="contact-page">
+	<div class="col-sm-12">
+  		<h2>Questions? Need Help?</h2>
+      <p>To make sure your question is answered as quickly as possible, please take a look at our <a href="#">FAQ</a> and <a href="#">documentation</a> first, before sending an inquiry directly to our staff using the form below.</p>
+	</div>
+</div>
+<!--/stories-->